# API_SWAGGER.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Traffic Management System API",
    "description": "Comprehensive REST API for intelligent traffic management and monitoring",
    "version": "1.0.0",
    "contact": {
      "name": "Support Team",
      "email": "support@traffic-system.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.traffic-system.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/detections": {
      "get": {
        "tags": ["Detections"],
        "summary": "List Detections",
        "description": "Get paginated list of vehicle detections",
        "operationId": "list_detections",
        "parameters": [
          {
            "name": "lane",
            "in": "query",
            "description": "Filter by lane",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "Start time filter",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "End time filter",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated detections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/detections/{detection_id}": {
      "get": {
        "tags": ["Detections"],
        "summary": "Get Detection Details",
        "description": "Get detailed information about a specific vehicle detection",
        "operationId": "get_detection",
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "description": "Detection ID",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detection details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDetection"
                }
              }
            }
          },
          "404": {
            "description": "Detection not found"
          }
        }
      }
    },
...
    "/signals/{lane}": {
      "get": {
        "tags": ["Signals"],
        "summary": "Get Signal State",
        "description": "Get current state of a specific traffic signal",
        "operationId": "get_signal_state",
        "parameters": [
          {
            "name": "lane",
            "in": "path",
            "description": "Lane name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/signals/{lane}/control": {
      "post": {
        "tags": ["Signals"],
        "summary": "Control Signal",
        "description": "Manually control a specific traffic signal",
        "operationId": "control_signal",
        "parameters": [
          {
            "name": "lane",
            "in": "path",
            "description": "Lane name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "Desired signal state",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SignalState"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "description": "Duration in seconds",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Control action confirmed"
          }
        }
      }
    },
...
    "/analytics/hourly": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Hourly Statistics",
        "description": "Get hourly traffic statistics",
        "operationId": "get_hourly_statistics",
        "parameters": [
          {
            "name": "lane",
            "in": "query",
            "description": "Filter by lane",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "description": "Number of hours to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 24,
              "minimum": 1,
              "maximum": 168
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hourly statistics"
          }
        }
      }
    },
    "/violations": {
      "get": {
        "tags": ["Violations"],
        "summary": "List Violations",
        "description": "Get paginated list of traffic violations",
        "operationId": "list_violations",
        "parameters": [
          {
            "name": "violation_type",
            "in": "query",
            "description": "Filter by type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ViolationType"
            }
          },
          {
            "name": "processed",
            "in": "query",
            "description": "Filter by processed status",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "min_severity",
            "in": "query",
            "description": "Minimum severity level",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1,
              "maximum": 5
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated violations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/violations/{violation_id}/process": {
      "post": {
        "tags": ["Violations"],
        "summary": "Mark Violation as Processed",
        "description": "Mark a violation record as processed",
        "operationId": "process_violation",
        "parameters": [
          {
            "name": "violation_id",
            "in": "path",
            "description": "Violation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "notes",
            "in": "query",
            "description": "Processing notes",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Violation processed"
          }
        }
      }
    },
    "/config/signals": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Get Signal Configuration",
        "description": "Get current signal timing configuration",
        "operationId": "get_signal_config",
        "responses": {
          "200": {
            "description": "Signal configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalTimingConfig"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Configuration"],
        "summary": "Update Signal Configuration",
        "description": "Update signal timing configuration",
        "operationId": "update_signal_config",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalTimingConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalTimingConfig"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SignalState": {
        "enum": ["green", "yellow", "red"],
        "type": "string",
        "title": "SignalState"
      },
      "ViolationType": {
        "enum": ["red_light", "speeding", "illegal_turn", "other"],
        "type": "string",
        "title": "ViolationType"
      },
      "VehicleClass": {
        "enum": ["car", "truck", "motorcycle", "bus", "other"],
        "type": "string",
        "title": "VehicleClass"
      },
...
      "VehicleDetection": {
        "type": "object",
        "required": ["detection_id", "timestamp", "lane", "vehicle_class", "confidence", "bounding_box"],
        "properties": {
          "detection_id": {
            "type": "integer",
            "title": "Detection ID"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "lane": {
            "type": "string",
            "title": "Lane"
          },
          "vehicle_class": {
            "$ref": "#/components/schemas/VehicleClass"
          },
          "confidence": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "title": "Confidence"
          },
          "bounding_box": {
            "type": "object",
            "title": "Bounding Box"
          },
          "speed": {
            "type": "number",
            "title": "Speed"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle ID"
          }
        }
      },
      "SignalStateResponse": {
        "type": "object",
        "required": ["lane", "state", "time_remaining", "cycle_position"],
        "properties": {
          "lane": {
            "type": "string",
            "title": "Lane"
          },
          "state": {
            "$ref": "#/components/schemas/SignalState"
          },
          "time_remaining": {
            "type": "integer",
            "title": "Time Remaining"
          },
          "cycle_position": {
            "type": "integer",
            "title": "Cycle Position"
          }
        }
      },
...
      "PageResponse": {
        "type": "object",
        "required": ["items", "total", "page", "page_size", "total_pages"],
        "properties": {
          "items": {
            "type": "array",
            "items": {},
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        }
      }
    }
  }
}